+++ packages/api/lib/controllers/boxesController.js
@@ -71,7 +71,7 @@ const
  * @apiParam (RequestBody) {MqttOption} [mqtt] settings for the MQTT integration of this senseBox
  * @apiParam (RequestBody) {TTNOption} [ttn] settings for the TTN integration of this senseBox
  * @apiParam (RequestBody) {String} [description] the updated description of this senseBox. Send '' (empty string) to delete this property.
- * @apiParam (RequestBody) {String} [image] the updated image of this senseBox encoded as base64 data uri.
+ * @apiParam (RequestBody) {String} [image] the updated image of this senseBox encoded as base64 data uri. To delete the current image, send 'deleteImage: true'.
  * @apiParam (RequestBody) {Object} [addons] allows to add addons to the box. Submit as Object with key `add` and the desired addon as value like `{"add":"feinstaub"}`
  * @apiParamExample {json} Request-Example:
  * {